requirements.txt
aiogram==2.22.2
loguru==0.6.0
PyYAML==6.0
